+++ src/app-index.ts
@@ -15,9 +15,9 @@ import globalStyleSheets, {init} from './global.css';
 import en from './translation/en';
 import fa from './translation/fa';
 
-import './elements/page-home';
-import './elements/page-game';
-import './elements/page-about';
+import './pages/page-home';
+import './pages/page-game';
+import './pages/page-about';
 
 import type {RoutesConfig} from '@alwatr/router';
 import type {ListenerInterface} from '@alwatr/signal';
